+++ pdf/test/all_tests.dart
@@ -29,6 +29,7 @@ import 'type1_test.dart' as type1;
 import 'widget_basic_test.dart' as widget_basic;
 import 'widget_clip_test.dart' as widget_clip;
 import 'widget_container_test.dart' as widget_container;
+import 'widget_flex_test.dart' as widget_flex;
 import 'widget_multipage_test.dart' as widget_multipage;
 import 'widget_table_test.dart' as widget_table;
 import 'widget_test.dart' as widget;
@@ -51,6 +52,7 @@ void main() {
   widget_basic.main();
   widget_clip.main();
   widget_container.main();
+  widget_flex.main();
   widget_multipage.main();
   widget_table.main();
   widget_text.main();